requirements.txt
discord
PyNaCl
art
python-dotenv